Atualizando as configurações do banco, criando o database, atualizando dependencias, atualizando o model e criando a classe UsuarioRepository
