+++ sistema-controle-doacao/src/main/java/org/sdp/Piex.java
@@ -1,0 +1,12 @@
+package org.sdp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class Piex {
+    public static void main(String[] args) {
+        SpringApplication.run(Piex.class, args);
+    }
+}
