+++ sistema-controle-doacao/src/main/java/org/sdp/domain/e/ETipoDoacao.java
@@ -1,4 +1,4 @@
-package org.sdp.model.e;
+package org.sdp.domain.e;
 
 public enum ETipoDoacao {
     Produtos, Dinheiro
