+++ sistema-controle-doacao/src/main/java/org/sdp/domain/produto/Produto.java
@@ -1,0 +1,25 @@
+package org.sdp.domain.produto;
+
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Table(name = "produtos")
+@Entity(name = "Produto")
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(of = "id")
+public class Produto {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String nome;
+    private int email;
+    private double valorProduto;
+
+
+}
