+++ sistema-controle-doacao/src/main/java/org/sdp/domain/usuario/UsuarioRepository.java
@@ -1,0 +1,8 @@
+package org.sdp.domain.usuario;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+
+}
