--- sistema-controle-doacao/src/main/java/org/sdp/model/Usuario.java
@@ -1,40 +1,0 @@
-package org.sdp.model;
-
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "tbUsuario")
-public class Usuario {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String login;
-    private String senha;
-
-    public Usuario(){
-
-    }
-    public Usuario(String login, String senha) {
-        this.login = login;
-        this.senha = senha;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getSenha() {
-        return senha;
-    }
-
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-
-}
