--- sistema-controle-doacao/src/main/java/org/sdp/util/JPAUtil.java
@@ -1,15 +1,0 @@
-package org.sdp.util;
-
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
-
-public class JPAUtil {
-
-    private static final EntityManagerFactory FACTORY = Persistence.createEntityManagerFactory("banco");
-
-    public static EntityManager getEntityManager(){
-        return FACTORY.createEntityManager();
-    }
-}
