+++ sistema-controle-doacao/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.datasource.url= jdbc:mysql://localhost:/projetopiex
+spring.datasource.username=root
+spring.datasource.password=aluno
+
+server.error.include-stacktrace=never
